Move toke dashboard types to common folder
