+++ src/common/store/pages/token-dashboard/slice.ts
@@ -2,14 +2,13 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 
 import { Chain } from 'common/models/Chain'
 import { BNWei, StringWei, WalletAddress } from 'common/models/Wallet'
-import { Nullable } from 'common/utils/typeUtils'
-
 import {
   AssociatedWallets,
   ConfirmRemoveWalletAction,
   ModalState,
   TokenDashboardState
-} from '../../../../store/token-dashboard/types'
+} from 'common/store/pages/token-dashboard/types'
+import { Nullable } from 'common/utils/typeUtils'
 
 const initialState: TokenDashboardState = {
   modalState: null,
